+++ src/grinch/__init__.py
@@ -1,4 +1,5 @@
 from .aliases import ADK, OBS, OBSM, UNS, VAR, VARM, AnnDataKeys
+from .base_processor import BaseProcessor
 from .conf import BaseConfig, BaseConfigurable
 from .de import TTest
 from .filters import FilterCells, FilterGenes
@@ -8,7 +9,6 @@ from .main import instantiate_config
 from .normalizers import Log1P, NormalizeTotal
 from .pipeline import GRPipeline
 from .predictors import BasePredictor, KMeans, LogisticRegression
-from .processors import BaseProcessor
 from .reporter import Report, Reporter
 from .splitter import DataSplitter, Splitter
 from .transformers import PCA, UMAP, BaseTransformer, TruncatedSVD
