+++ src/grinch/base_processor.py
