+++ src/grinch/indexer.py
@@ -3,7 +3,7 @@ from anndata import AnnData
 from pydantic import validator
 from sklearn.utils import column_or_1d
 
-from .processors import BaseProcessor
+from .base_processor import BaseProcessor
 from .utils.validation import validate_axis
