+++ src/grinch/splitter.py
@@ -6,8 +6,8 @@ from anndata import AnnData
 from pydantic import Field, validate_arguments, validator
 from sklearn.model_selection import train_test_split
 
+from .base_processor import BaseProcessor
 from .conf import BaseConfigurable
-from .processors import BaseProcessor
 from .utils.validation import all_not_None, any_not_None
