+++ src/grinch/transformers.py
@@ -9,7 +9,7 @@ from sklearn.decomposition import TruncatedSVD as _TruncatedSVD
 from umap import UMAP as _UMAP
 
 from .aliases import OBSM, UNS
-from .processors import BaseProcessor
+from .base_processor import BaseProcessor
 from .utils.validation import check_has_processor, pop_args
 
 logger = logging.getLogger(__name__)
